requirements.txt
asgiref==3.8.1
bleach==6.1.0
Django==5.1.2
django-filter==24.3
django-rest-knox==5.0.2
djangorestframework==3.15.2
drf-api-logger==1.1.16
mysqlclient==2.2.5
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
webencodings==0.5.1
